Fixes for MacOS 15.4
On 15.4, it seems that /private got more restrictive with SIP, and disallows setting a whole bunch of metadata. Since we end up just creating a /private/etc/puppetlabs with nothing in it, this just uses -k flag instead of -ka with rsync. Also, since we usually don't end up putting anything in /var, allow that not to be present in the tarball.

Additionally, the version of CMake in 15.4 does not allow the minimum CMake version to be declared in a Makefile to be anything less than 3.5, and some of them have 3.4 declared. This adds a flag to override what's found in the Makefile to declare 3.5 to be the minimum. It doesn't matter that much because the version of CMake we are using is 4.x.